+++ docs/source/releasenotes.rst
@@ -9,6 +9,22 @@ Release notes
 `Released <https://pypi.org/project/rpaframework/#history>`_
 ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
+22.5.2 - 13 Apr 2023
+--------------------
+
+- Library **RPA.Browser.Selenium** (:pr:`916`):
+
+  - Takes into use the latest Selenium (**4.8.3** at the time of the release) and its
+    logging fixes. (:issue:`883`)
+  - Fixes scheme ensuring bug during not provided URLs. (:issue:`586`)
+  - Fixes `headless` mode setting with newer browsers and enables more browser features
+    when running in this mode. (:issue:`919`)
+  - Ensures keyword ``Screenshot`` works and logs messages as expected. (logs aren't
+    polluted with raw image data on levels equal or greater than `INFO` and it returns
+    the final output image path; :issue:`920`)
+  - Takes into use the latest webdriver-manager fixes given ``rpaframework-core``
+    **11.0.1**.
+
 22.5.1 - 07 Apr 2023
 --------------------